Split runnable method from LevelTypeSelectView for use in project
