+++ src/builderMockups/Main.java
@@ -1,0 +1,22 @@
+package builderMockups;
+
+import java.awt.EventQueue;
+
+import javax.swing.UIManager;
+
+public class Main {
+
+	public static void main(String[] args) {
+		EventQueue.invokeLater(new Runnable() {
+			public void run() {
+				try {
+					UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
+					LevelTypeSelectView window = new LevelTypeSelectView();
+					window.setVisible(true);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+			}
+		});
+	}
+}
